+++ articles/urls.py
@@ -1,8 +1,14 @@
 from django.conf.urls import url
 
 from . import views
+from . import autocomplete_views as autocomplete
 
 
 urlpatterns = [
+    url(
+        r'^article-autocomplete/$',
+        autocomplete.ArticleAutocomplete.as_view(),
+        name='article-autocomplete',
+    ),
     url(r'^(?P<slug>[^/]+)/$', views.ArticleDetailView.as_view(), name='detail'),
 ]